pyproject.toml
[tool.poetry]
name = "photogrid"
version = "0.1.1"
description = ""
authors = ["Marco Rougeth <pypi@rougeth.com>"]

[tool.poetry.dependencies]
python = "^3.9"
Pillow = "^8.4.0"
click = "^8.0.3"
rich = "^10.12.0"

[tool.poetry.dev-dependencies]

[tool.poetry.scripts]
photogrid = "photogrid:cli"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
